+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## Unreleased
 
 - Major: Add notifier for when items are bought or sold on the Grand Exchange. (#275)
+- Minor: Fire level notification upon reaching max skill experience of 200M. (#273)
+- Minor: Include clan name in notificiation metadata. (#274)
+- Dev: Update gradle wrapper to v8.2.1 patch version. (#276)
 
 ## 1.5.3